Feature/ganon as py module (#134)
* split classes files ganon python

* ship scripts on setup.py

* better code for split parts

* config with argparser, need to update tests

* better file organization, basic test passing

* fixes and trigger travis

* fix unit test

* make install

* return error without params, install travis test

* sudo on make install

* tests travis install

* entry point ganon python

* fix modules python

* fix modules tests

* travis all builds, remove warnings

* remove release from install

* test osx sudo

* activate all tests
