+++ tests/ganon/integration/test_classify.py
@@ -1,7 +1,7 @@
-import unittest
-import shlex, pickle
+import unittest, shlex, pickle, sys
 from pathlib import Path
-from src.ganon import ganon
+sys.path.append('src')
+from ganon import ganon
 
 class TestClassify(unittest.TestCase):
     def test_classify(self):
@@ -10,8 +10,13 @@ class TestClassify(unittest.TestCase):
         """
         path_data = "tests/ganon/integration/data/"
         prefix = "test_classify"
-        ret = ganon.main(shlex.split("./ganon classify --output-all --db-prefix "+path_data+"sample_bacteria --single-reads "+path_data+"bacteria.simulated.1.fq -o "+prefix))
-        
+
+        ret = ganon.main("classify", 
+                        output_all=True, 
+                        db_prefix=path_data+"sample_bacteria",
+                        single_reads=path_data+"bacteria.simulated.1.fq",
+                        output_prefix=prefix)
+
         # check if ran okay
         self.assertFalse(ret, "ganon classify finish with an error")
