+++ tests/ganon/integration/test_report.py
@@ -1,7 +1,7 @@
-import unittest
-import shlex, pickle
+import unittest, shlex, pickle, sys
 from pathlib import Path
-from src.ganon import ganon
+sys.path.append('src')
+from ganon import ganon
 
 class TestReport(unittest.TestCase):
     def test_report(self):
@@ -10,8 +10,13 @@ class TestReport(unittest.TestCase):
         """
         path_data = "tests/ganon/integration/data/"
         output_file = "test_report.tre"
-        ret = ganon.main(shlex.split("./ganon report --output-report "+output_file+" --db-prefix "+path_data+"sample_bacteria --rep-file "+path_data+"results.rep --rank all"))
-        
+
+        ret = ganon.main("report", 
+                        output_report=output_file, 
+                        db_prefix=path_data+"sample_bacteria",
+                        rep_file=path_data+"results.rep",
+                        rank="all")
+
         # check if ran okay
         self.assertFalse(ret, "ganon report finish with an error")
